refactor(dgq): move the decode-only subset into crates/dgemm::format
The GEMM crate now owns the decode side of every weight format the engine
stores, so a fused body in dgemm can read the engine's packed weights without
depending on the engine:

  format/layout.rs  block constants and every *_row_bytes / *_matrix_bytes
  format/bf16.rs    bf16_to_f32 (identical to src/shaders/cpu/mod.rs), bits/bytes helpers
  format/fp4.rs     E2M1 + FP8-E4M3 codecs (moved whole, with its 5 tests)
  format/block.rs   q4/q6/q8 decode, *_weight_at, q4/q8/q6_gemm_cpu
  format/nvfp4.rs   nvfp4 decode, nvfp4_weight_at, nvfp4_gemm_cpu

The engine keeps what is not decode: the manifest and classification, every
quantize_* encoder, store/overlay/hf_resolve/spot_check/embed_row, and the
manifest-shaped dequant_to_f32 glue. Each engine dgq module re-exports the
moved names, so NOT ONE engine call site changed: crate::dgq::block::
dequant_matrix_q4 and crate::dgq::layout::q4_row_bytes still resolve.

gpukit::Error gains Format(&'static str) and the engine's From maps it back to
Error::Format, so error text and variants are unchanged (that impl also loses
its macOS-only gate, which was already wrong once non-GPU code can produce
one).

Identity evidence: all 26 moved functions were extracted from git HEAD and
diffed against the new location -- byte-identical except the two spec'd
pub(crate) -> pub signatures; format/fp4.rs is byte-identical whole.

Gates (engine gate is failure-set identity, see below): engine 756 passed /
25 failed / 42 ignored, failing set IDENTICAL to the pre-change baseline;
dgemm 15/15 and dgops 50/50 on Metal AND on CUDA (GB10, sm_121); clippy clean
for dgemm/gpukit/diffgemma; fmt clean; nanogpt forward parity unchanged at
max_abs 2.384e-7.

The 25 engine failures are environmental, not from this change: the q4 pack is
an interrupted download (model.dgq.json present, model.dgq.bin a 0-byte
.incomplete blob), so the heavy model-gated tests panic ENOENT instead of
skipping. dgq_model_dir() should gate on the blob, not just the manifest
(PLAN.md). The 761 -> 756 pass delta is exactly the 5 fp4 tests that moved
into dgemm.
