+++ Cargo.toml
@@ -43,6 +43,7 @@ scattered-collect = "0.22"
 opt-level = 1
 
 [target.'cfg(target_os = "macos")'.dependencies]
+dgemm = { path = "crates/dgemm" }
 gpukit = { path = "crates/gpukit" }
 objc2 = { version = "0.6", default-features = false, features = ["std"] }
 objc2-metal = { version = "0.3", default-features = false, features = [
